+++ src/Menu/Menu.Component.js
@@ -27,48 +27,17 @@ export const MenuComponent = () => {
             </Example>
 
             <Example
-                description='Menu with grouped sub-menus and group headers.'
-                title='Menu with Group'>
-                <Menu>
-                    <Menu.List>
-                        <Menu.Item>
-                            <Link to='#'>Option 1</Link>
-                        </Menu.Item>
-                        <Menu.Item>
-                            <Link to='#'>Option 2</Link>
-                        </Menu.Item>
-                        <Menu.Item>
-                            <Link to='#'>Option 3</Link>
-                        </Menu.Item>
-                    </Menu.List>
-                    <Menu.Group title='Group Header'>
-                        <Menu.List>
-                            <Menu.Item>
-                                <Link to='#'>Option 4</Link>
-                            </Menu.Item>
-                            <Menu.Item>
-                                <Link to='#'>Option 5</Link>
-                            </Menu.Item>
-                            <Menu.Item>
-                                <Link to='#'>Option 6</Link>
-                            </Menu.Item>
-                        </Menu.List>
-                    </Menu.Group>
-                </Menu>
-            </Example>
-
-            <Example
                 description='Menu items with horizontal line as separator.'
                 title='Menu with Separator'>
                 <Menu>
-                    <Menu.List separated>
-                        <Menu.Item>
+                    <Menu.List>
+                        <Menu.Item separator>
                             <Link to='#'>Option 1</Link>
                         </Menu.Item>
                         <Menu.Item>
                             <Link to='#'>Option 2</Link>
                         </Menu.Item>
-                        <Menu.Item>
+                        <Menu.Item separator>
                             <Link to='#'>Option 3</Link>
                         </Menu.Item>
                         <Menu.Item>
@@ -79,20 +48,41 @@ export const MenuComponent = () => {
             </Example>
 
             <Example
-                description='Menu items with an add-on before.'
-                title='Menu with Addon Before'>
-                <Menu addonBefore>
+                description='Menu item states - selected, active, and disabled.'
+                title='Menu with Separator'>
+                <Menu>
                     <Menu.List>
-                        <Menu.Item>
+                        <Menu.Item separator>
                             <Link to='#'>Option 1</Link>
                         </Menu.Item>
-                        <Menu.Item addon='accept'>
+                        <Menu.Item active>
                             <Link to='#'>Option 2</Link>
                         </Menu.Item>
-                        <Menu.Item>
+                        <Menu.Item selected>
                             <Link to='#'>Option 3</Link>
                         </Menu.Item>
-                        <Menu.Item>
+                        <Menu.Item disabled>
+                            <Link to='#'>Option 4</Link>
+                        </Menu.Item>
+                    </Menu.List>
+                </Menu>
+            </Example>
+
+            <Example
+                description='Menu items with an add-on before or after.'
+                title='Menu with Add-on'>
+                <Menu>
+                    <Menu.List>
+                        <Menu.Item addonBefore='grid'>
+                            <Link to='#'>Option 1</Link>
+                        </Menu.Item>
+                        <Menu.Item addonBefore='wrench'>
+                            <Link to='#'>Option 2</Link>
+                        </Menu.Item>
+                        <Menu.Item addonAfter='accept' addonBefore='history'>
+                            <Link to='#'>Option 3</Link>
+                        </Menu.Item>
+                        <Menu.Item addonBefore='lightbulb'>
                             <Link to='#'>Option 4</Link>
                         </Menu.Item>
                     </Menu.List>