+++ src/SearchInput/SearchInput.js
@@ -32,7 +32,15 @@ class SearchInput extends Component {
     };
 
     handleListItemClick = (event, item) => {
-        item.callback ? item.callback() : null;
+        this.setState((prevState) => ({
+            value: item.text,
+            isExpanded: false,
+            searchExpanded: false,
+            filteredResult: this.filterList(prevState.searchList, item.text)
+        }), () => {
+            item?.callback();
+            this.props.onSelect(event, item);
+        });
     };
 
     handleChange = event => {
@@ -104,6 +112,7 @@ class SearchInput extends Component {
             onEnter,
             searchList,
             onChange,
+            onSelect,
             noSearchBtn,
             compact,
             className,
@@ -133,19 +142,18 @@ class SearchInput extends Component {
                     {this.state.filteredResult && this.state.filteredResult.length > 0 ? (
                         this.state.filteredResult.map((item, index) => {
                             return (
-                                <li
-                                    className='fd-menu__item'
+                                <Menu.Item
                                     key={index}
                                     onClick={(e) => this.handleListItemClick(e, item)}>
                                     <strong>{this.state.value}</strong>
                                     {this.state.value && this.state.value.length
                                         ? item.text.substring(this.state.value.length)
                                         : item.text}
-                                </li>
+                                </Menu.Item>
                             );
                         })
                     ) : (
-                        <li className='fd-menu__item'>No result</li>
+                        <Menu.Item>No result</Menu.Item>
                     )}
                 </Menu.List>
             </Menu>
@@ -197,6 +205,7 @@ class SearchInput extends Component {
                     disableStyles={disableStyles}
                     noArrow
                     onClickOutside={this.handleClickOutside}
+                    show={this.state.isExpanded}
                     widthSizingType='matchTarget' />
             </div>
         );
@@ -229,12 +238,14 @@ SearchInput.propTypes = {
         text: PropTypes.string
     }),
     onChange: PropTypes.func,
-    onEnter: PropTypes.func
+    onEnter: PropTypes.func,
+    onSelect: PropTypes.func
 };
 
 SearchInput.defaultProps = {
     onChange: () => { },
-    onEnter: () => { }
+    onEnter: () => { },
+    onSelect: () => { }
 };
 
 SearchInput.propDescriptions = {